Add form to create new report.
This functionality may change in the future.

This is mainly for testing purposes. No idea if this functionality will
be available (from the web) at the final release of the service. Maybe
users should be allowed to sent reports both from their android devices and
the web, but the ones created from the web form should pass some kind of
validation first.